+++ sort/selection_sort.h
@@ -5,10 +5,10 @@
  * @version 0.1
  * @date 2019-05-09
  */
+
 #ifndef SORT_SELECTION_SORT_H_
 #define SORT_SELECTION_SORT_H_
 
-#include <cstddef>
 #include <utility>
 
 namespace algo {
@@ -21,11 +21,11 @@ namespace algo {
  * @param size Size of the input array, i.e., size of nums.
  */
 template<typename ValueType>
-void SelectionSort(ValueType nums[], const size_t size) {
-  size_t min_idx;
-  for (size_t i = 0; i < size; ++i) {
+void SelectionSort(ValueType nums[], const int size) {
+  int min_idx;
+  for (int i = 0; i < size; ++i) {
     min_idx = i;
-    for (size_t j = i + 1; j < size; ++j) {
+    for (int j = i + 1; j < size; ++j) {
       if (nums[j] < nums[min_idx]) {
         min_idx = j;
       }
